+++ Changelog.md
@@ -1,4 +1,20 @@
-# Changelog v1.7.1
+# Changelog v1.8
 
-## Fixes
-- Shearing mossy blocks causes game crash
+## Additions
+- Added blocks
+  - Added "Crimson Moss"
+  - Added "Crimson Moss Carpet"
+  - Added "Warped Moss"
+  - Added "Warped Moss Carpet"
+- Added items
+  - Added "Crimson Moss"
+  - Added "Crimson Moss Carpet"
+  - Added "Crimson Moss Ball"   
+  - Added "Warped Moss"
+  - Added "Warped Moss Carpet"
+  - Added "Warped Moss Ball" 
+
+## Changes
+- Added new moss blocks to:
+  - "Crimson Forest"
+  - "Warped Forest"